Fix temporal RDO biasing distortion
Before this commit, temporal RDO scaled distortion by:

    const FACTOR: f32 = 3.;
    const ADDEND: f64 = 0.65;
    (mean_importance / FACTOR) as f64 + ADDEND;

For an unreferenced block, `mean_importance` is 0, which means that in
intra frames we were systematically scaling the distortion by 0.65. This
is bad since it's equivalent to picking the wrong lambda for RDO.

This commit fixes this by replacing the temporal RDO scaling formula by
one based on the original mbtree paper which was designed to have no
effect on unreferenced frames.