Fix desync when using multiple segments
The segmentation index is a partition-level flag, but in
`encode_tx_block` we were looking it up using the offset of the current
transform block. This _almost_ works but can be wrong for chroma blocks
because the transform block offset might not be contained in the
partition (see added documentation for an example). We could fix this
locally in `encode_tx_block` by passing `tile_partition_bo` instead of
`tx_bo` to `get_qidx`, but it's simpler to just have `qidx` be a
parameter of `encode_tx_block` since the caller already knows what the
correct `qidx` is anyway.

This commit only reenables segmentation at speed 0 since it heavily
impacts speed right now, see FIXME comment.